fossid-webapp: Rename the "api" package to "model"
The term "model" is more common in ORT for packages containing data model
classes.
