--- clients/fossid-webapp/src/main/kotlin/api/summary/License.kt
@@ -1,28 +1,0 @@
-/*
- * Copyright (C) 2020 Bosch.IO GmbH
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- * SPDX-License-Identifier: Apache-2.0
- * License-Filename: LICENSE
- */
-
-package org.ossreviewtoolkit.clients.fossid.api.summary
-
-import org.ossreviewtoolkit.clients.fossid.api.identification.common.LicenseMatchType
-
-data class License(
-    val identifier: String,
-    val name: String?,
-    val origin: LicenseMatchType
-)
