+++ clients/fossid-webapp/src/main/kotlin/model/identification/identifiedFiles/IdentifiedFile.kt
@@ -17,25 +17,30 @@
  * License-Filename: LICENSE
  */
 
-package org.ossreviewtoolkit.clients.fossid.api.identification.markedAsIdentified
+package org.ossreviewtoolkit.clients.fossid.model.identification.identifiedFiles
 
 import com.fasterxml.jackson.annotation.JsonUnwrapped
 
-import org.ossreviewtoolkit.clients.fossid.api.identification.common.Component
-import org.ossreviewtoolkit.clients.fossid.api.summary.License
-import org.ossreviewtoolkit.clients.fossid.api.summary.Summarizable
-import org.ossreviewtoolkit.clients.fossid.api.summary.SummaryIdentifiedFile
+import org.ossreviewtoolkit.clients.fossid.model.identification.common.Component
+import org.ossreviewtoolkit.clients.fossid.model.summary.License
+import org.ossreviewtoolkit.clients.fossid.model.summary.Summarizable
+import org.ossreviewtoolkit.clients.fossid.model.summary.SummaryIdentifiedFile
 
-data class MarkedAsIdentifiedFile(
+data class IdentifiedFile(
     val comment: String?,
 
-    val identificationId: Int,
-
+    val identificationId: Int?,
     val identificationCopyright: String,
 
     val isDistributed: Int,
 
-    val rowId: Int
+    val rowId: Int,
+
+    val userName: String?,
+
+    val userSurname: String?,
+
+    val userUsername: String?,
 ) : Summarizable {
     @JsonUnwrapped(prefix = "component_")
     lateinit var component: Component // TODO: How to get the additional components? Only one is returned
@@ -47,9 +52,9 @@ data class MarkedAsIdentifiedFile(
         val licenses = file.licenses?.let { licenses ->
             licenses.values.map {
                 License(
-                    identifier = it.file.licenseIdentifier!!,
-                    name = it.file.licenseName,
-                    origin = it.type
+                    identifier = it.identifier,
+                    name = it.name,
+                    origin = it.fileLicenseMatchType
                 )
             }
         }.orEmpty()
